Update region :background variable
Changed the region :background variable from everforest-hard-dark-gray
to everforest-hard-dark-gutter. The previous gray color, used for
highlighting the current line, was causing issues in visual mode within
evil mode, making it difficult to distinguish wether the line is
selected or highlighted. By switching the background color of the region
to gutter, we enhance the contrast and provide a clearer indication of
the currently selected lines.